Improve request pipeline
Improve requests stability.
Task pool is now much easier to use.
Update of the oneshots window.
The oneShot window and speeches work in the same way.
Requests will now be preserved even when there is a reload of scripts or an editor crash.
It is now possible to make several oneShot requests at the same time.
The oneShot generation no longer leaves any clips on the API.